Add updated naming system for clarity while in development
