+++ client/src/components/NavBar/NavBar.scss
@@ -50,6 +50,7 @@
 .burger-menu-links {
     height: 100%;
     display: flex;
+    white-space: nowrap;
 
     .navbar-link,
     .burger-menu-link {
