+++ client/src/constants/userLevels.js
@@ -1,4 +1,4 @@
-export const levels = [
+export const userLevels = [
     0, 10, 40, 100, 175, 275, 400, 550, 725, 900, 1100, 1440, 2000, 2750, 3750,
     5000, 6500, 8250, 10000, 12200, 15300, 21400, 28900, 39100, 53400, 68800,
     87600, 110600, 150000, 200000, 300000,
