+++ client/src/pages/Battle/matchFunctions.jsx
