+++ client/src/pages/BattleSetup/BattleSetup.scss
