--- client/src/pages/MatchEnd/MatchEnd.scss
@@ -1,161 +1,0 @@
-.end {
-    align-items: center;
-    justify-content: center;
-    color: var(--white-color);
-
-    .result {
-        font-size: 160px;
-        text-shadow: 4px 4px 1px var(--black-color),
-            -2px -2px 1px var(--black-color), 2px -2px 1px var(--black-color),
-            -2px 2px 1px var(--black-color);
-
-        @media screen and (max-width: 900px) {
-            font-size: 124px;
-        }
-
-        @media screen and (max-width: 600px) {
-            font-size: 96px;
-        }
-    }
-
-    .rewards {
-        width: 75%;
-        display: flex;
-        align-items: center;
-        justify-content: space-evenly;
-
-        .reward {
-            height: 200px;
-            width: 25%;
-            display: flex;
-            flex-direction: column;
-            align-items: center;
-            justify-content: center;
-            background-color: var(--background-color);
-            border: 1px solid var(--primary-color);
-            border-radius: 8px;
-            font-size: 32px;
-
-            @media screen and (max-width: 900px) {
-                height: 100px;
-                width: 35%;
-                font-size: 24px;
-            }
-
-            @media screen and (max-width: 600px) {
-                width: 40%;
-            }
-
-            span {
-                display: flex;
-                align-items: center;
-                justify-content: center;
-            }
-
-            img {
-                width: 15%;
-                margin-left: 8px;
-            }
-        }
-    }
-
-    .results {
-        position: relative;
-        height: 400px;
-        width: 80vw;
-        display: flex;
-        align-items: center;
-        justify-content: space-around;
-
-        @media screen and (max-width: 600px) {
-            flex-direction: column;
-        }
-
-        .one,
-        .two {
-            display: flex;
-            flex: 1;
-            align-items: center;
-            justify-content: space-evenly;
-
-            @media screen and (max-width: 600px) {
-                width: 100%;
-                flex-direction: row;
-            }
-        }
-
-        .one {
-            @media screen and (max-width: 600px) {
-                flex-direction: row;
-            }
-        }
-
-        .two {
-            @media screen and (max-width: 600px) {
-                flex-direction: row-reverse;
-            }
-        }
-
-        .player {
-            display: flex;
-            flex-direction: column;
-            align-items: center;
-            justify-content: center;
-            gap: 24px;
-
-            @media screen and (max-width: 600px) {
-                flex-direction: row;
-            }
-
-            h1 {
-                margin-top: 0;
-            }
-
-            &__image {
-                width: 10vw;
-                border: 1px solid var(--primary-color);
-                border-radius: 8px;
-            }
-        }
-
-        .score {
-            font-size: 124px;
-
-            @media screen and (max-width: 1200px) {
-                font-size: 96px;
-            }
-
-            @media screen and (max-width: 900px) {
-                font-size: 72px;
-            }
-
-            @media screen and (max-width: 600px) {
-                font-size: 56px;
-            }
-        }
-
-        .buttons {
-            display: flex;
-            flex-direction: column;
-            gap: 48px;
-
-            @media screen and (max-width: 900px) {
-                gap: 36px;
-            }
-
-            .button {
-                position: static;
-
-                @media screen and (max-width: 900px) {
-                    padding: 8px;
-                    font-size: 20px;
-                }
-
-                @media screen and (max-width: 600px) {
-                    padding: 4px 0;
-                    font-size: 16px;
-                }
-            }
-        }
-    }
-}
