+++ client/src/pages/index.js
@@ -1,10 +1,8 @@
 export { default as Account } from './Account/Account'
 export { default as Collection } from './Collection/Collection'
 export { default as Home } from './Home/Home'
-export { default as LandingPage } from './LandingPage/LandingPage'
+export { default as Landing } from './Landing/Landing'
 export { default as Market } from './Market/Market'
-export { default as Match } from './Match/Match'
-export { default as MatchEnd } from './MatchEnd/MatchEnd'
-export { default as MatchSetup } from './MatchSetup/MatchSetup'
-export { default as Discovery } from './Discovery/Discovery'
 export { default as Battle } from './Battle/Battle'
+export { default as BattleSetup } from './BattleSetup/BattleSetup'
+export { default as OpenPacks } from './OpenPacks/OpenPacks'
